+++ packages/typescript/ai-anthropic/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @tanstack/ai-anthropic
 
+## 0.2.0
+
+### Minor Changes
+
+- Standard schema / standard json schema support for TanStack AI ([#165](https://github.com/TanStack/ai/pull/165))
+
+### Patch Changes
+
+- Updated dependencies [[`c5df33c`](https://github.com/TanStack/ai/commit/c5df33c2d3e72c3332048ffe7c64a553e5ea86fb)]:
+  - @tanstack/ai@0.2.0
+
+## 0.1.0
+
+### Minor Changes
+
+- Split up adapters for better tree shaking into separate functionalities ([#137](https://github.com/TanStack/ai/pull/137))
+
+### Patch Changes
+
+- Updated dependencies [[`8d77614`](https://github.com/TanStack/ai/commit/8d776146f94ffd1579e1ab01b26dcb94d1bb3092)]:
+  - @tanstack/ai@0.1.0
+
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies [[`52c3172`](https://github.com/TanStack/ai/commit/52c317244294a75b0c7f5e6cafc8583fbb6abfb7)]:
+  - @tanstack/ai@0.0.3
+
 ## 0.0.2
 
 ### Patch Changes