+++ packages/typescript/react-ai-devtools/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @tanstack/react-ai-devtools
 
+## 0.1.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/ai-devtools-core@0.1.1
+
+## 0.1.0
+
+### Minor Changes
+
+- Split up adapters for better tree shaking into separate functionalities ([#137](https://github.com/TanStack/ai/pull/137))
+
+### Patch Changes
+
+- Updated dependencies [[`8d77614`](https://github.com/TanStack/ai/commit/8d776146f94ffd1579e1ab01b26dcb94d1bb3092)]:
+  - @tanstack/ai-devtools-core@0.1.0
+
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies [[`52c3172`](https://github.com/TanStack/ai/commit/52c317244294a75b0c7f5e6cafc8583fbb6abfb7)]:
+  - @tanstack/ai-devtools-core@0.0.3
+
 ## 0.0.2
 
 ### Patch Changes