+++ packages/typescript/smoke-tests/adapters/CHANGELOG.md
@@ -1,5 +1,49 @@
 # @tanstack/tests-adapters
 
+## 0.1.0
+
+### Minor Changes
+
+- Standard schema / standard json schema support for TanStack AI ([#165](https://github.com/TanStack/ai/pull/165))
+
+### Patch Changes
+
+- Updated dependencies [[`c5df33c`](https://github.com/TanStack/ai/commit/c5df33c2d3e72c3332048ffe7c64a553e5ea86fb)]:
+  - @tanstack/ai-anthropic@1.0.0
+  - @tanstack/ai-ollama@1.0.0
+  - @tanstack/ai@0.2.0
+  - @tanstack/ai-gemini@1.0.0
+  - @tanstack/ai-openai@1.0.0
+
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies [[`2615fd7`](https://github.com/TanStack/ai/commit/2615fd7552c77888db970cb9dd1d64f73257e96c)]:
+  - @tanstack/ai-openai@0.1.1
+
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies [[`8d77614`](https://github.com/TanStack/ai/commit/8d776146f94ffd1579e1ab01b26dcb94d1bb3092)]:
+  - @tanstack/ai-anthropic@0.1.0
+  - @tanstack/ai-gemini@0.1.0
+  - @tanstack/ai-ollama@0.1.0
+  - @tanstack/ai-openai@0.1.0
+  - @tanstack/ai@0.1.0
+
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies [[`78f15fe`](https://github.com/TanStack/ai/commit/78f15fea74c7bfd5fcdb0fe106968b110f0b2bc2), [`52c3172`](https://github.com/TanStack/ai/commit/52c317244294a75b0c7f5e6cafc8583fbb6abfb7)]:
+  - @tanstack/ai-openai@0.0.3
+  - @tanstack/ai@0.0.3
+  - @tanstack/ai-anthropic@0.0.3
+  - @tanstack/ai-gemini@0.0.3
+  - @tanstack/ai-ollama@0.0.3
+
 ## 0.0.2
 
 ### Patch Changes